main: reset zoom view when change to Victory scene + restructure project files
