+++ include/core/level/level.h
@@ -10,7 +10,7 @@
 #include <memory>
 #include "tile.h"
 #include "vec2.h"
-#include "common_constants.h"
+#include "core/engine/common_constants.h"
 
 enum class ObjectType
 {
