+++ include/core/manager/texture_manager.h
