+++ include/scene/gameplay_scene/system/ray_appender_system.h
@@ -5,7 +5,7 @@
 
 #include "system.h"
 #include "entity_manager.h"
-#include "level_manager.h"
+#include "core/manager/level_manager.h"
 
 /**
  * [System Description]
