+++ src/core/action.cpp
@@ -1,4 +1,4 @@
-#include "action.h"
+#include "core/engine/action.h"
 
 Action::Action(Action::Type type, Action::Mode mode) : m_type(type), m_mode(mode)
 {
