+++ src/core/texture_manager.cpp
@@ -1,4 +1,4 @@
-#include "texture_manager.h"
+#include "core/manager/texture_manager.h"
 
 TextureManager::TextureManager() = default;
